+++ logic/job/comet_info.go
@@ -9,18 +9,18 @@ const (
 )
 
 var (
-	RoomServersMap = make(map[int32]map[int32]struct{}) // roomid:servers
+	RoomServersMap = make(map[int64]map[int32]struct{}) // roomid:servers
 )
 
 func MergeRoomServers() {
 	var (
 		c           *Comet
 		ok          bool
-		roomId      int32
+		roomId      int64
 		serverId    int32
-		roomIds     map[int32]struct{}
+		roomIds     map[int64]struct{}
 		servers     map[int32]struct{}
-		roomServers = make(map[int32]map[int32]struct{})
+		roomServers = make(map[int64]map[int32]struct{})
 	)
 	// all comet nodes
 	for serverId, c = range cometServiceMap {
